math_parser: remove diag func. declarations from header
